refactor: replaced gnav with rounded nav bar

+++ lib/src/presentation/viewmodel/folder_page_viewmodel.dart
@@ -9,17 +9,25 @@ class FolderPageViewmodel extends AsyncNotifier<List<Folder>> {
     final folders = foldermaps
         .map((map) => Folder(id: map["id"], name: map["name"]))
         .toList();
-    return folders;
+
+    final defaultFolder = Folder(id: "default", name: "General");
+    return [defaultFolder, ...folders];
   }
 
   Future<void> addFolder({required String name, required String id}) async {
     final newFolder = Folder(id: id, name: name);
+    if (newFolder.id == "default" || newFolder.name == "General") {
+      return;
+    }
     await DatabaseHelper.instance.insertFolder(id: id, name: name);
 
     state = AsyncValue.data([...state.value!, newFolder]);
   }
 
   Future<void> deleteFolder(Folder folder) async {
+    if (folder.id == "default") {
+      return;
+    }
     await DatabaseHelper.instance.deleteFolderById(folder.id);
     state = AsyncValue.data(
       state.value!.where((f) => f.id != folder.id).toList(),
